+++ User_Space/Polaris_Network_Implementation.h
@@ -13,8 +13,8 @@ namespace Network_Components
 		{
 			struct
 			{
-				long id;
-				long dir;
+				int id;
+				int dir;
 			};
 			long long id_dir;
 		};
